+++ internal/cmd/grpc.go
@@ -249,8 +249,8 @@ func NewGRPCServer(
 			cacheShutdown errFunc
 			err           error
 		)
-
 		cacher, cacheShutdown, err = getCache(ctx, cfg)
+
 		if err != nil {
 			return nil, err
 		}
@@ -315,7 +315,7 @@ func NewGRPCServer(
 
 	// cache must come after auth interceptors
 	if cfg.Cache.Enabled && cacher != nil {
-		interceptors = append(interceptors, middlewaregrpc.CacheControlUnaryInterceptor, middlewaregrpc.EvaluationCacheUnaryInterceptor(cacher, logger))
+		interceptors = append(interceptors, middlewaregrpc.CacheUnaryInterceptor(cacher, logger))
 	}
 
 	// audit sinks configuration