+++ ui/src/App.tsx
@@ -23,6 +23,7 @@ const Segments = loadable(() => import('./app/segments/Segments'));
 const Console = loadable(() => import('./app/console/Console'));
 const Login = loadable(() => import('./app/auth/Login'));
 const Settings = loadable(() => import('./app/Settings'));
+const Support = loadable(() => import('./app/Support'));
 const Preferences = loadable(() => import('./app/preferences/Preferences'));
 const Namespaces = loadable(() => import('./app/namespaces/Namespaces'));
 const Tokens = loadable(() => import('./app/tokens/Tokens'));
@@ -116,6 +117,10 @@ const router = createHashRouter([
             element: <Tokens />
           }
         ]
+      },
+      {
+        path: 'support',
+        element: <Support />
       },
       ...namespacesRoutes
     ]