+++ CHANGELOG.md
@@ -3,6 +3,30 @@
 This format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.26.0](https://github.com/flipt-io/flipt/releases/tag/v1.26.0) - 2023-08-28
+
+### Added
+
+- exp: Homebrew tap (#2044)
+- Switch to monaco editor for console (#2040)
+- default to ~/Library/Application\ Support/flipt for db sqlite file on mac (#2039)
+- `auth`: Add parameters for pkce flow
+- `auth`: OAuth for GitHub (#2016)
+
+### Changed
+
+- Dependency updates
+
+### Fixed
+
+- hover icon visibility for copy key to clipboard in lightmode (#2056)
+
+## [v1.25.2](https://github.com/flipt-io/flipt/releases/tag/v1.25.2) - 2023-08-23
+
+### Fixed
+
+- `internal/ext`: add 1.1 to versions and validate explicit 1.2 (#2042)
+
 ## [v1.25.1](https://github.com/flipt-io/flipt/releases/tag/v1.25.1) - 2023-08-21
 
 ### Fixed